@@ -1,43 +1,43 @@
-/*
- * Copyright (c) 2018-2020 C4
- *
- * This file is part of Curios, a mod made for Minecraft.
- *
- * Curios is free software: you can redistribute it and/or modify it
- * under the terms of the GNU Lesser General Public License as published
- * by the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * Curios is distributed in the hope that it will be useful, but
- * WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR PARTICULAR PURPOSE.  See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with Curios.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-package top.theillusivec4.curios.common.integration.jei;
-
-import javax.annotation.Nonnull;
-import mezz.jei.api.IModPlugin;
-import mezz.jei.api.JeiPlugin;
-import mezz.jei.api.registration.IGuiHandlerRegistration;
-import net.minecraft.resources.ResourceLocation;
-import top.theillusivec4.curios.Curios;
-import top.theillusivec4.curios.client.gui.CuriosScreen;
-
-@JeiPlugin
-public class CuriosJeiPlugin implements IModPlugin {
-
-  @Override
-  @Nonnull
-  public ResourceLocation getPluginUid() {
-    return new ResourceLocation(Curios.MODID, Curios.MODID);
-  }
-
-  @Override
-  public void registerGuiHandlers(IGuiHandlerRegistration registration) {
-    registration.addGuiContainerHandler(CuriosScreen.class, new CuriosContainerHandler());
-  }
-}
+///*
+// * Copyright (c) 2018-2020 C4
+// *
+// * This file is part of Curios, a mod made for Minecraft.
+// *
+// * Curios is free software: you can redistribute it and/or modify it
+// * under the terms of the GNU Lesser General Public License as published
+// * by the Free Software Foundation, either version 3 of the License, or
+// * (at your option) any later version.
+// *
+// * Curios is distributed in the hope that it will be useful, but
+// * WITHOUT ANY WARRANTY; without even the implied warranty of
+// * MERCHANTABILITY or FITNESS FOR PARTICULAR PURPOSE.  See the
+// * GNU Lesser General Public License for more details.
+// *
+// * You should have received a copy of the GNU Lesser General Public
+// * License along with Curios.  If not, see <https://www.gnu.org/licenses/>.
+// */
+//
+//package top.theillusivec4.curios.common.integration.jei;
+//
+//import javax.annotation.Nonnull;
+//import mezz.jei.api.IModPlugin;
+//import mezz.jei.api.JeiPlugin;
+//import mezz.jei.api.registration.IGuiHandlerRegistration;
+//import net.minecraft.resources.ResourceLocation;
+//import top.theillusivec4.curios.Curios;
+//import top.theillusivec4.curios.client.gui.CuriosScreen;
+//
+//@JeiPlugin
+//public class CuriosJeiPlugin implements IModPlugin {
+//
+//  @Override
+//  @Nonnull
+//  public ResourceLocation getPluginUid() {
+//    return new ResourceLocation(Curios.MODID, Curios.MODID);
+//  }
+//
+//  @Override
+//  public void registerGuiHandlers(IGuiHandlerRegistration registration) {
+//    registration.addGuiContainerHandler(CuriosScreen.class, new CuriosContainerHandler());
+//  }
+//}
