@@ -30,7 +30,7 @@ import java.util.TreeMap;
 import javax.annotation.Nonnull;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraftforge.fmllegacy.network.PacketDistributor;
+import net.minecraftforge.network.PacketDistributor;
 import top.theillusivec4.curios.api.CuriosApi;
 import top.theillusivec4.curios.api.type.ISlotType;
 import top.theillusivec4.curios.api.type.inventory.ICurioStacksHandler;
@@ -61,10 +61,15 @@ public class SlotHelper implements ISlotHelper {
   }
 
   @Override
+  public Collection<ISlotType> getSlotTypes(LivingEntity livingEntity) {
+    return getSlotTypes();
+  }
+
+  @Override
   public SortedMap<ISlotType, ICurioStacksHandler> createSlots() {
     SortedMap<ISlotType, ICurioStacksHandler> curios = new TreeMap<>();
     this.getSlotTypes().forEach(type -> curios.put(type,
-        new CurioStacksHandler(null, type.getIdentifier(), type.getSize(), 0, type.isVisible(),
+        new CurioStacksHandler(null, type.getIdentifier(), type.getSize(), type.isVisible(),
             type.hasCosmetic())));
     return curios;
   }
@@ -74,8 +79,8 @@ public class SlotHelper implements ISlotHelper {
     SortedMap<ISlotType, ICurioStacksHandler> curios = new TreeMap<>();
     CuriosApi.getCuriosHelper().getCuriosHandler(livingEntity).ifPresent(
         handler -> this.getSlotTypes().forEach(type -> curios.put(type,
-            new CurioStacksHandler(handler, type.getIdentifier(), type.getSize(), 0,
-                type.isVisible(), type.hasCosmetic()))));
+            new CurioStacksHandler(handler, type.getIdentifier(), type.getSize(), type.isVisible(),
+                type.hasCosmetic()))));
     return curios;
   }
 
